refactor(enrollment): collapse ParticipantPolicy to schema-as-struct
Slice 2 of the Enrollment flatten. Merge domain model + Ecto schema +
mapper + repository + 2 ports into KlassHero.Enrollment.ParticipantPolicy;
inline set/get/eligibility use cases onto the context. The eligibility
check calls the ParticipantDetails/ProgramSchedule ACLs directly (their
single-impl ports go in the final slice). set_participant_policy keeps its
participant_policy_set event dispatch, wrapped in context_span.

- Dropped the dead get_policies_by_program_ids batch read (no callers).
- CreateEnrollment eligibility gate repointed to the public API.
- Test tree flattened: participant_policy_test + check_participant_eligibility_test
  replace the deleted model/schema/mapper/repo/command/query unit tests.
