+++ config/config.exs
@@ -20,7 +20,6 @@ alias KlassHero.Enrollment.Adapters.Driven.ACL.ProgramScheduleACL
 alias KlassHero.Enrollment.Adapters.Driven.Notifications.InviteEmailNotifier
 alias KlassHero.Enrollment.Adapters.Driven.Persistence.Repositories.BulkEnrollmentInviteRepository
 alias KlassHero.Enrollment.Adapters.Driven.Persistence.Repositories.EnrollmentRepository
-alias KlassHero.Enrollment.Adapters.Driven.Persistence.Repositories.ParticipantPolicyRepository
 alias KlassHero.Enrollment.Adapters.Driving.Events.InviteFamilyReadyHandler
 alias KlassHero.Family.Adapters.Driving.Events.FamilyEventHandler
 alias KlassHero.Family.Adapters.Driving.Events.InviteClaimedHandler
@@ -187,8 +186,6 @@ config :klass_hero, :default_tz, "Europe/Berlin"
 config :klass_hero, :enrollment,
   for_managing_enrollments: EnrollmentRepository,
   for_querying_enrollments: EnrollmentRepository,
-  for_managing_participant_policies: ParticipantPolicyRepository,
-  for_querying_participant_policies: ParticipantPolicyRepository,
   for_resolving_participant_details: ParticipantDetailsACL,
   for_resolving_program_schedule: ProgramScheduleACL,
   for_resolving_child_info: ChildInfoACL,
